Replace size_t with int
Replace size_t with int for loop indices, matrix accessors, and related members to unify signed index types and avoid unsigned/signed conversions. Changes include: Ant constructor parameter, Ant methods and loops, Matrix operator() and constructors, CsrMatrix offsets type and row access, and multiple loops in AntColony. This enables using -1 as a sentinel (e.g. unassigned schedule) and keeps index types consistent across the solver implementation.

+++ solver/include/ant.hpp
@@ -46,7 +46,7 @@ public:
   /**
    * @brief Initializes an ant with empty schedule and default tracking matrices. 
    */
-  Ant(size_t n_exams, size_t n_slots);
+  Ant(int n_exams, int n_slots);
 
   /**
    * @brief Overloads the less-than operator for Ant comparison.
